+++ lib/crawler/fetcher/recorder.ex
@@ -14,7 +14,7 @@ defmodule Crawler.Fetcher.Recorder do
       {:ok, %{depth: 3, url: "url1"}}
 
       iex> Recorder.record(url: "url2", depth: 2)
-      iex> Store.find("url2")
+      iex> Store.find({"url2", nil})
       %Page{url: "url2"}
   """
   def record(opts) do
@@ -35,19 +35,19 @@ defmodule Crawler.Fetcher.Recorder do
       {:ok, nil}
 
       iex> Recorder.record(url: "url", depth: 2)
-      iex> Recorder.maybe_store_page("body", %{store: Store, url: "url"})
-      {:ok, {%Page{url: "url", body: "body", opts: %{store: Store, url: "url"}}, %Page{url: "url", body: nil}}}
+      iex> Recorder.maybe_store_page("body", %{store: Store, url: "url", scope: nil})
+      {:ok, {%Page{url: "url", body: "body", opts: %{store: Store, url: "url", scope: nil}}, %Page{url: "url", body: nil}}}
   """
   def maybe_store_page(_body, %{store: nil} = _opts) do
     {:ok, nil}
   end
 
   def maybe_store_page(body, opts) do
-    {:ok, opts[:store].add_page_data(opts[:url], body, opts)}
+    {:ok, opts[:store].add_page_data({opts[:url], opts[:scope]}, body, opts)}
   end
 
   defp store_url(opts) do
-    Store.add(opts[:url])
+    Store.add({opts[:url], opts[:scope]})
   end
 
   defp store_url_depth(opts) do