+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## master
 
+- [Added] Add `:force` option
+- [Added] Add `:scope` option
+
 ## v1.4.0 [2023-10-07]
 
 - [Added] Allow multiple instances of Crawler sharing the same queue